# lxdapi/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "魔方财务系统专用LXD容器管理后端API服务",
        "title": "魔方财务-LXD对接API",
        "termsOfService": "https://github.com/xkatld/zjmf-lxd-server",
        "contact": {
            "name": "API支持",
            "url": "https://github.com/xkatld/zjmf-lxd-server",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/xkatld/zjmf-lxd-server/blob/main/LICENSE"
        },
        "version": "1.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/ipv6/add": {
            "post": {
                "description": "为容器添加独立的公网IPv6地址绑定",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPv6绑定"
                ],
                "summary": "添加IPv6独立绑定",
                "parameters": [
                    {
                        "type": "string",
                        "description": "容器名称",
                        "name": "hostname",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "绑定描述",
                        "name": "description",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.IPv6BindingResponse"
                        }
                    }
                }
            }
        },
        "/api/ipv6/delete": {
            "post": {
                "description": "删除容器的独立公网IPv6地址绑定",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPv6绑定"
                ],
                "summary": "删除IPv6独立绑定",
                "parameters": [
                    {
                        "type": "string",
                        "description": "容器名称",
                        "name": "hostname",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "公网IPv6地址",
                        "name": "public_ipv6",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.IPv6BindingResponse"
                        }
                    }
                }
            }
        },
        "/api/ipv6/list": {
            "get": {
                "description": "获取指定容器的IPv6绑定列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPv6绑定"
                ],
                "summary": "获取IPv6绑定列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "容器名称",
                        "name": "hostname",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.IPv6BindingListResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.IPv6BindingListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.IPv6BindingRule"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "models.IPv6BindingResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "models.IPv6BindingRule": {
            "type": "object",
            "properties": {
                "container_ipv6": {
                    "description": "容器内网IPv6地址",
                    "type": "string"
                },
                "container_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "interface": {
                    "description": "绑定的网卡接口",
                    "type": "string"
                },
                "public_ipv6": {
                    "description": "公网IPv6地址",
                    "type": "string"
                },
                "status": {
                    "description": "active, inactive, error",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "在header中添加apikey字段，值为配置文件中的api_hash",
            "type": "apiKey",
            "name": "apikey",
            "in": "header"
        }
    }
}